Upravy pred bufferovanymi triedami.

+++ WebCrawler/src/main/java/Jsoup_crawler/Caller.java
@@ -1,21 +1,15 @@
 package Jsoup_crawler;
 
-
-import java.io.BufferedReader;
-import java.io.FileReader;
-import java.io.IOException;
-
-//trieda na ukladanie ano
 class Caller {
-    private int deep;
+    private long deep;
     private String url;
 
-    public Caller(int deep, String url) {
+    public Caller(long deep, String url) {
         this.deep = deep;
         this.url = url;
     }
 
-    public int deep(){
+    public long deep(){
         return this.deep;
     }
